safari-600.7-branch/LayoutTests/plugins/netscape-plugin-map-data-to-src-expected.txt

CONSOLE MESSAGE: PLUGIN: src: blah

This tests that an object tag with a data attribute gets a src attribute as well.
